% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cfl.R
\docType{data}
\name{cfl}
\alias{cfl}
\title{Characteristics of 50 small cities in central Florida}
\format{A data frame with 50 rows and 16 variables:
\describe{
  \item{name}{Name of city}
  \item{hh_inc}{Median household income}
  \item{pc_inc}{Per capita income}
  \item{age}{Median age}
  \item{pop2000}{Population in 2000}
  \item{area}{Area in square miles}
  \item{density}{Population density per square mile}
  \item{col}{Cost of Living index}
  \item{pop2014}{Population in 2014}
  \item{popchange}{Percent change in population from 2000 to 2014}
  \item{hh_size}{Median household size}
  \item{h_price}{Median house price}
  \item{vcrime}{Violent crime rate}
  \item{pcrime}{Property crime rate}
  \item{lat}{latitude}
  \item{long}{longitude}
}}
\source{
Miscellaneous. Mostly \url{http://www.wikipedia.com/}
}
\usage{
cfl
}
\description{
A dataset containing the locations and characteristics of 50 small cities in
central Florida.
}
\keyword{datasets}

