% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/citySelector.R
\name{citySelector}
\alias{citySelector}
\title{citySelector}
\usage{
citySelector()
}
\value{
A named vector of city scores
}
\description{
citySelector launches a shiny gadget, a shiny app that returns a value when
closed. The app displays the towns in the cfl data set (50 small towns in
Central Florida) and lets a user build a scoring function to score the towns
in an interactive, gui fashion. To close the app and retrieve the scores for
each city, click the "Get Scores" button.
}

