% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mapVis.R
\name{intervalScale}
\alias{intervalScale}
\title{intervalScale}
\usage{
intervalScale(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
A numeric vector
}
\description{
Scales a vector to the 0-1 interval. The minimum value will be scaled to
zero, the maximum to 1, and the remaining values will be spaced
proportionately between.
}
\examples{
intervalScale(-5:5)
}

