% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mapVis.R
\name{mapVis}
\alias{mapVis}
\title{mapVis}
\usage{
mapVis(x, radius = ~10, color = "Blue")
}
\arguments{
\item{x}{A data frame with \code{lat} (latitude) and \code{long} (longitude)
columns.}

\item{radius}{Expression to map circle marker radius to. Can use column names
from \code{x}.}

\item{color}{Color of circle markers, as character sting}
}
\value{
A leaflet map visualization
}
\description{
Plots an interactive leaflet map that displays each row as a circle marker in
the map. Designed to work with the \code{\link{cfl}} data set.
}

